switch CitiesReducer to ts
