+++ src/components/cities-list/cities-list.tsx
@@ -1,7 +1,7 @@
 import React, { Dispatch } from 'react';
 import {connect} from 'react-redux';
 import ActionCreator from '../../store/action-creator/action-creator';
-import {getCities} from '../../store/reducers/citiesReducer/selectors';
+import {getCities} from '../../store/reducers/cities-reducer/selectors';
 import {getCurrentCity} from '../../store/reducers/currentCityReducer/selectors';
 import {ICity, IStore} from '../../types/types';
