+++ src/store/reducers/cities-reducer/cities-reducer.test.ts
@@ -1,9 +1,9 @@
 import cities from '../../../mock/test-cities';
-import citiesReducer from './citiesReducer';
+import citiesReducer from './cities-reducer';
 
 describe(`citiesReducer`, () => {
   it(`should return first argumet`, () => {
-    const result = citiesReducer(cities, {type: `SOME_ACTION`, cities: []});
+    const result = citiesReducer(cities);
     expect(result).toEqual(cities);
   });
 });
