+++ src/store/reducers/cities-reducer/selectors.ts
@@ -1,0 +1,6 @@
+import NameSpace from '../name-space';
+import {IStore} from '../../../types/types';
+
+export const getCities = (state: IStore) => {
+  return state[NameSpace.CITIES];
+};
