--- src/store/reducers/citiesReducer/selectors.js
@@ -1,5 +1,0 @@
-import NameSpace from '../name-space';
-
-export const getCities = (state) => {
-  return state[NameSpace.CITIES];
-};
